requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-requests==0.1.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cchardet==2.1.7
cffi==1.15.1
charset-normalizer==2.1.0
creart==0.2.1
creart-graia==0.1.5
croniter==1.3.5
frozenlist==1.3.0
fuzzywuzzy==0.18.0
graia-amnesia==0.5.3
graia-ariadne==0.7.15
graia-broadcast==0.18.2
graia-saya==0.0.16
graia-scheduler==0.0.8
idna==3.3
launart==0.5.0
loguru==0.6.0
lxml==4.9.1
multidict==6.0.2
Naked==0.1.31
numpy==1.23.0
prompt-toolkit==3.0.30
pycares==4.2.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
six==1.16.0
statv==0.2.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
yarl==1.7.2
